chore: move interfaces to proper file
